Refactor fatzebra library code into a standard Magento model, remove dead code (non-tokenization/refund code)
